move duplicate serialization logic to base class
