+++ ee/vellum_ee/workflows/display/nodes/base_node_display.py
@@ -247,6 +247,31 @@ class BaseNodeDisplay(Generic[NodeType], metaclass=BaseNodeDisplayMeta):
 
         return ports
 
+    def _serialize_attributes(self, display_context: "WorkflowDisplayContext"):
+        """Serialize node attributes, skipping unserializable ones."""
+        attributes = []
+        for attribute in self._node:
+            if attribute in self.__unserializable_attributes__:
+                continue
+
+            id = (
+                str(self.attribute_ids_by_name[attribute.name])
+                if self.attribute_ids_by_name.get(attribute.name)
+                else str(uuid4_from_hash(f"{self.node_id}|{attribute.name}"))
+            )
+            try:
+                attributes.append(
+                    {
+                        "id": id,
+                        "name": attribute.name,
+                        "value": serialize_value(display_context, attribute.instance),
+                    }
+                )
+            except ValueError as e:
+                raise ValueError(f"Failed to serialize attribute '{attribute.name}': {e}")
+
+        return attributes
+
     def get_base(self) -> CodeResourceDefinition:
         node = self._node
